Remove `initialize` from MapReduceForm.
This CL moves MapReduceForm from being 1:1 with an IterativeProcess to being
1:1 with a Computation.

We want to continue to be able to evaluate if computations are compatible
conceptually with single federated rounds. However, we also need our data
structures to acknowledge that some federated executions require different
pairings of computations (2 or more). By separating the "initialization" part
of the MapReduceForm from the content that defines the single federated round,
we will have more flexibility when defining the executions we want.

This CL continues optionally to allow IterativeProcesses to be translated into
MapReduceForm, but the initialize portion, after being validated and
simplified, is kept separate.

PiperOrigin-RevId: 461942785